+++ src/application/messaging/command/user/handler/user_register_command_handler.py
@@ -3,10 +3,11 @@ from application.messaging.command.user.user_register_command import (
     UserRegisterCommand,
     UserRegisterCommandResult,
 )
+from domain.auth.auth_info.local.value_objects import PlainPassword
 from domain.auth.auth_info.value_objects import HashedPassword
 from domain.user.repository.user_repository import UserRepository
 from domain.user.user import User
-from domain.user.value_objects import Email, PlainPassword, Username
+from domain.user.value_objects import Email, Username
 from shared_kernel.hasher.hasher import Hasher
 from shared_kernel.time.time_provider import TimeProvider
