+++ tests/unit/domain/auth/auth_info/local/test_value_objects.py
@@ -1,7 +1,6 @@
 import pytest
 
-from domain.auth.auth_info.local.value_objects import PlainPassword
-from domain.user.exceptions import (
+from domain.auth.auth_info.local.exceptions import (
     PasswordMissingLowercaseError,
     PasswordMissingNumberError,
     PasswordMissingSpecialCharacterError,
@@ -9,6 +8,7 @@ from domain.user.exceptions import (
     PasswordTooLongError,
     PasswordTooShortError,
 )
+from domain.auth.auth_info.local.value_objects import PlainPassword
 
 
 class TestPlainPasswordVO:
